put SideBar in a single component and fix the bugs
