+++ src/components/SideBar.jsx
@@ -1,194 +1,131 @@
-// import List from "@mui/material/List";
-// import Divider from "@mui/material/Divider";
-// import IconButton from "@mui/material/IconButton";
-// import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
-// import ChevronRightIcon from "@mui/icons-material/ChevronRight";
-// import ListItem from "@mui/material/ListItem";
-// import ListItemButton from "@mui/material/ListItemButton";
-// import ListItemIcon from "@mui/material/ListItemIcon";
-// import ListItemText from "@mui/material/ListItemText";
-// import InboxIcon from "@mui/icons-material/MoveToInbox";
-// import MailIcon from "@mui/icons-material/Mail";
-// import { styled, useTheme } from "styled-components";
-// import MuiDrawer from "@mui/material/Drawer";
+import List from "@mui/material/List";
+import Divider from "@mui/material/Divider";
+import IconButton from "@mui/material/IconButton";
+import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
+import ChevronRightIcon from "@mui/icons-material/ChevronRight";
+import ListItem from "@mui/material/ListItem";
+import ListItemButton from "@mui/material/ListItemButton";
+import ListItemIcon from "@mui/material/ListItemIcon";
+import ListItemText from "@mui/material/ListItemText";
+import InboxIcon from "@mui/icons-material/MoveToInbox";
+import MailIcon from "@mui/icons-material/Mail";
+import { styled, useTheme } from "@mui/material/styles";
+import MuiDrawer from "@mui/material/Drawer";
 
-// const drawerWidth = 240;
+const drawerWidth = 240;
 
-// const openedMixin = (theme) => ({
-//   width: drawerWidth,
-//   transition: theme.transitions.create("width", {
-//     easing: theme.transitions.easing.sharp,
-//     duration: theme.transitions.duration.enteringScreen,
-//   }),
-//   overflowX: "hidden",
-// });
+const openedMixin = (theme) => ({
+  width: drawerWidth,
+  transition: theme.transitions.create("width", {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.enteringScreen,
+  }),
+  overflowX: "hidden",
+});
 
-// const closedMixin = (theme) => ({
-//   transition: theme.transitions.create("width", {
-//     easing: theme.transitions.easing.sharp,
-//     duration: theme.transitions.duration.leavingScreen,
-//   }),
-//   overflowX: "hidden",
-//   width: `calc(${theme.spacing(7)} + 1px)`,
-//   [theme.breakpoints.up("sm")]: {
-//     width: `calc(${theme.spacing(8)} + 1px)`,
-//   },
-// });
+const closedMixin = (theme) => ({
+  transition: theme.transitions.create("width", {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.leavingScreen,
+  }),
+  overflowX: "hidden",
+  width: `calc(${theme.spacing(7)} + 1px)`,
+  [theme.breakpoints.up("sm")]: {
+    width: `calc(${theme.spacing(9)} + 1px)`,
+  },
+});
 
-// const Drawer = styled(MuiDrawer, {
-//   shouldForwardProp: (prop) => prop !== "open",
-// })(({ theme }) => ({
-//   width: drawerWidth,
-//   flexShrink: 0,
-//   whiteSpace: "nowrap",
-//   boxSizing: "border-box",
-//   variants: [
-//     {
-//       props: ({ open }) => open,
-//       style: {
-//         ...openedMixin(theme),
-//         "& .MuiDrawer-paper": openedMixin(theme),
-//       },
-//     },
-//     {
-//       props: ({ open }) => !open,
-//       style: {
-//         ...closedMixin(theme),
-//         "& .MuiDrawer-paper": closedMixin(theme),
-//       },
-//     },
-//   ],
-// }));
+const Drawer = styled(MuiDrawer, {
+  shouldForwardProp: (prop) => prop !== "open",
+// @ts-ignore
+})(({ theme, open }) => ({
+  width: drawerWidth,
+  flexShrink: 0,
+  whiteSpace: "nowrap",
+  boxSizing: "border-box",
+  ...(open && {
+    ...openedMixin(theme),
+    "& .MuiDrawer-paper": openedMixin(theme),
+  }),
+  ...(!open && {
+    ...closedMixin(theme),
+    "& .MuiDrawer-paper": closedMixin(theme),
+  }),
+}));
 
-// const DrawerHeader = styled("div")(({ theme }) => ({
-//   display: "flex",
-//   alignItems: "center",
-//   justifyContent: "flex-end",
-//   padding: theme.spacing(0, 1),
-//   // necessary for content to be below app bar
-//   ...theme.mixins.toolbar,
-// }));
+const DrawerHeader = styled("div")(({ theme }) => ({
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "flex-end",
+  padding: theme.spacing(0, 1),
+  ...theme.mixins.toolbar,
+}));
 
-// // eslint-disable-next-line react/prop-types
-// export default function SideBar({ open, handleDrawerClose }) {
-//   const theme = useTheme();
-//   return (
-//     <div>
-//       <Drawer variant="permanent" open={open}>
-//         <DrawerHeader>
-//           <IconButton onClick={handleDrawerClose}>
-//             {theme.direction === "rtl" ? (
-//               <ChevronRightIcon />
-//             ) : (
-//               <ChevronLeftIcon />
-//             )}
-//           </IconButton>
-//         </DrawerHeader>
-//         <Divider />
-//         <List>
-//           {["Inbox", "Starred", "Send email", "Drafts"].map((text, index) => (
-//             <ListItem key={text} disablePadding sx={{ display: "block" }}>
-//               <ListItemButton
-//                 sx={[
-//                   {
-//                     minHeight: 48,
-//                     px: 2.5,
-//                   },
-//                   open
-//                     ? {
-//                         justifyContent: "initial",
-//                       }
-//                     : {
-//                         justifyContent: "center",
-//                       },
-//                 ]}
-//               >
-//                 <ListItemIcon
-//                   sx={[
-//                     {
-//                       minWidth: 0,
-//                       justifyContent: "center",
-//                     },
-//                     open
-//                       ? {
-//                           mr: 3,
-//                         }
-//                       : {
-//                           mr: "auto",
-//                         },
-//                   ]}
-//                 >
-//                   {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
-//                 </ListItemIcon>
-//                 <ListItemText
-//                   primary={text}
-//                   sx={[
-//                     open
-//                       ? {
-//                           opacity: 1,
-//                         }
-//                       : {
-//                           opacity: 0,
-//                         },
-//                   ]}
-//                 />
-//               </ListItemButton>
-//             </ListItem>
-//           ))}
-//         </List>
-//         <Divider />
-//         <List>
-//           {["All mail", "Trash", "Spam"].map((text, index) => (
-//             <ListItem key={text} disablePadding sx={{ display: "block" }}>
-//               <ListItemButton
-//                 sx={[
-//                   {
-//                     minHeight: 48,
-//                     px: 2.5,
-//                   },
-//                   open
-//                     ? {
-//                         justifyContent: "initial",
-//                       }
-//                     : {
-//                         justifyContent: "center",
-//                       },
-//                 ]}
-//               >
-//                 <ListItemIcon
-//                   sx={[
-//                     {
-//                       minWidth: 0,
-//                       justifyContent: "center",
-//                     },
-//                     open
-//                       ? {
-//                           mr: 3,
-//                         }
-//                       : {
-//                           mr: "auto",
-//                         },
-//                   ]}
-//                 >
-//                   {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
-//                 </ListItemIcon>
-//                 <ListItemText
-//                   primary={text}
-//                   sx={[
-//                     open
-//                       ? {
-//                           opacity: 1,
-//                         }
-//                       : {
-//                           opacity: 0,
-//                         },
-//                   ]}
-//                 />
-//               </ListItemButton>
-//             </ListItem>
-//           ))}
-//         </List>
-//       </Drawer>
-//     </div>
-//   );
-// }
+// eslint-disable-next-line react/prop-types
+export default function SideBar({ open, handleDrawerClose }) {
+  const theme = useTheme();
+
+  return (
+    <Drawer variant="permanent" open={open}>
+      <DrawerHeader>
+        <IconButton onClick={handleDrawerClose}>
+          {theme.direction === "rtl" ? (
+            <ChevronRightIcon />
+          ) : (
+            <ChevronLeftIcon />
+          )}
+        </IconButton>
+      </DrawerHeader>
+      <Divider />
+      <List>
+        {["Inbox", "Starred", "Send email", "Drafts"].map((text, index) => (
+          <ListItem key={text} disablePadding sx={{ display: "block" }}>
+            <ListItemButton
+              sx={{
+                minHeight: 48,
+                justifyContent: open ? "initial" : "center",
+                px: 2.5,
+              }}
+            >
+              <ListItemIcon
+                sx={{
+                  minWidth: 0,
+                  mr: open ? 3 : "auto",
+                  justifyContent: "center",
+                }}
+              >
+                {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
+              </ListItemIcon>
+              <ListItemText primary={text} sx={{ opacity: open ? 1 : 0 }} />
+            </ListItemButton>
+          </ListItem>
+        ))}
+      </List>
+      <Divider />
+      <List>
+        {["All mail", "Trash", "Spam"].map((text, index) => (
+          <ListItem key={text} disablePadding sx={{ display: "block" }}>
+            <ListItemButton
+              sx={{
+                minHeight: 48,
+                justifyContent: open ? "initial" : "center",
+                px: 2.5,
+              }}
+            >
+              <ListItemIcon
+                sx={{
+                  minWidth: 0,
+                  mr: open ? 3 : "auto",
+                  justifyContent: "center",
+                }}
+              >
+                {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
+              </ListItemIcon>
+              <ListItemText primary={text} sx={{ opacity: open ? 1 : 0 }} />
+            </ListItemButton>
+          </ListItem>
+        ))}
+      </List>
+    </Drawer>
+  );
+}
